Some plumbing, needs polishing
Some plumbing of expressions had been done
Now all expressions needs to be typed.
Function body type checking needs to be
done with unions. For example if a function
body has 3 lines producing int, float, int
and the codomain of the function is
int since (int | float | int) produces true
we can say at compile time that this function
might be valid, so we leave it alone. If
the body is something like (string, string, string)
we say this function is not valid.
There is no possibility that it can produce
int. We also check the type of return value during
the parsing and compare it with codomain of the
function