A new more formal type system
A new more formal type system is underway.

The type system is simply simply typed lambda calculus.
The citation of the book is in TypeContext.hs
The rules in the book should be implemented and the
constructs of the language should be revised to support
it fully. Parsing should have the following stages:

- Symbolic Syntax Tree: which is being done in ASTree.hs
module.
- Expansion of Symbolic syntax tree into full program tree.
This level is about macros and other meta programming related
stuff.
- Transformation of Expanded Tree -> Expression Tree: the expression
tree must consist of only two elements function definitions
and function applications. Literals are considered as
unnamed function definitions.
- Type checking the expression tree: The expression
tree is type checked using the simply typed lambda
calculus rules

Once the typechecking is done, we should evaluate the
expression tree.