Lots of new notes added probably notes are done
Most of the components of the new language is documented.
The macro system is going to be the most complicated
element of the system since it allows a lot of flexibility.
Macro expansion environments, functional macros with
type checking etc.

Abstraction, type system, and flow statements are more or
less done. Record and context are also done.
Flow binds now has a new syntax, with new operators.
Macros have also new operators that needs to be documented.

The next step is to update bnf grammar and start with
the implementation of lexer.

The bnf needs to take into account the array declaration
syntax, the macro operators, and bunch of other new
stuff that can be found in the notes and not in the
bnf representation and usage examples. The notes are
now the main reference.