New notes on records
Now module syntax is done. Abstraction syntax is almost done.
However records had changed significantly. I have decided to
let go of partial declarations, as there should be only one
way of doing things whenever possible. Same goes with
bindind. I might decide to use only partial declaration and
binding with compile errors later on, but for now, everything
is declared and bind in one go.
Current way of records are at least coherent.
Contexts are not. Without a semantic extension mechanism,
there is not a lot of things you can do to unify programs.
Contexts should work like typeclasses. However there are no
placeholder generic types in lambda script. So the type
of the abstraction that is going to be extended through
context needs to be prepended or appended to the type
declaration of the context abstractions.
Should context only work for records like typeclasses
that works only with ADTs ?