Binding syntax change
Binding syntax changed from := to =:
Now we have a clear delegation of operators. Those starting
with ":" is about declaring. Whether it is declaring
an abstraction, a record or a concept.
"=" is about binding. Again not really important what
we are binding. It concerns itself with binding stuff.
">" is about execution flow. It can be excluded or, or 'and'
execution flow which gets to be parallelized automatically.

The only thing that is lacking now is a coherent way of dealing
with macros.