Macro system more or less thought not done
The macro system more or less had been done.
However the matching function for coming up
with new syntax rules has not been done yet.
We are providing nice split merge facilities
for macro parameters to be used in macro body.
We should also provide a simple match facility
to do pattern matching of the code.