Macro system seems to be done now onto memory model
The memory model needs to be clarified. Should it be
manual management, which would be a little weird
since everything is immutable, but not unimaginable.
Should there be a gc ? I think it should be manual
but not sure how to deal with its semantics since
all values are read only. We can only signal
the destruction of certain abstractions and records
after some point in computation. What to do with
recursive functions ?

Once the memory model and its required operations
are figured out, we can start bnf grammar of the
language then start implementing it