The recursive and mutually recursive structures
can be supported through macros. The abstractions
however can not produce recursive code. This is a good
balance between flexibility and performance.
The non recursive nature of runtime code helps us
to evaluate it easily in GPU, and recursive nature
of compile time code helps us to use flexible structures