Macro system more or less done now memory model
Macro system of the language is more or less done. Now, we need to handle
how to deal with the memory model of the language.
Couple of ideas:
The syntax tree provides a state machine normally. However, if
we can obtain a task graph, that has a tree structure, that is
a graph with no back edges, it might be doable to use the structure
to favor manual memory management. However without a mutuable state,
there is no point in doing manual memory management. If all is going
to be immutable, we can do reference counting or some other thing
manage memory.
If we are going to implement mutuable abstractions.
We can do it by adding a suffix like # it would be something like
(:$ ) for immutable abstractions and (:$#) for mutuable
abstractions. The behaviour is that you can evoke bind operations
multiple times to mutuable abstractions. However this requires us to
think with time which was absent until here, so it might not be a
good idea to introduce it here.