Module bnf done, memory management still not specified
The bnf grammar for modules are done.
The memory management is still not specified. Since
it implies a lot about the structure of the programs,
I decided to read a little before deciding on anything.
There are couple of books that can checked out in the
next session. Memory management
with C/C++, the garbage collection handbook,
ASAP: As Static As Possible memory management.