Notes added with respect to core of lambdascript
The notes are added with respect to core of lambdascript
There is still the problem of orientation and sequence
computation. If an order is not specified for a given
block of computation, we treat the contents of the block
as an abstract binding tree where the each variable
is a node and calls determine the edges. The
practical programming languages book has an implementation
of abstract binding tree.
If however the sequence indicator is specified for the
block, we evaluate expressions one after the other.
Writting order is also important to think through